fix: seven input monsters - NaN's forever-dirty buffer, NUL's blind skip, the seq overflow's immortal job, the JS precision freeze, the depth freeze, the name cap, the TOAST drag (#521)
* fix: the attack-input monsters close at their mechanisms

The input monsters through the real path (enqueue -> dispatch -> progress
-> terminal -> admin render), what broke, and the gate each one gets:

- percent=NaN poisoned the coalesce buffer: the retire protocol keys on
  ==, and nan != nan, so one ctx.progress(percent=nan) kept the buffer
  dirty FOREVER - the flush loop re-wrote the row every tick for the
  job's lifetime. Finiteness gate at the caller-supplied door (range
  stays the actor's semantics: percent=150.0 stays legal).
- The wire-divergence monsters (percent inf / str, step str, detail
  non-str): encoders rewrote them between the two progress surfaces.
  Type-checked at the door, nothing mutates on a refusal.
- The NUL monster's missing rooms (#414): the progress door now refuses
  a NUL in detail and data on the bytes the size check already encoded;
  the flush guard stays as defense-in-depth. Previously the flush only
  SKIPPED the row and re-logged the same permanent defect every tick
  forever, the actor none the wiser.
- The size monster: detail was unbounded while data had
  progress_data_max_bytes. Same ceiling, same serialized-bytes
  measurement, same ProgressTooLarge, before the buffer mutates.
- The depth monster: orjson's encoder recursion limit is the door for
  deep data (pinned at depth 600 and 10_000, buffer untouched); the
  admin portal's fingerprint machine is depth-capped in realtime.js so
  no server state can RangeError the canonicalize or the stringify
  beneath it, and the render degrades instead of freezing the timeline.
- The name monsters: a 100k-char queue or actor name passed validation
  and failed the jobs INSERT in btree index-item arithmetic (incompress-
  ible names over ~2700 bytes, ProgramLimitExceededError, opaque at
  enqueue). Bounded at 255 characters at both chokepoints, the tag
  bound's precedent; the rejection echoes at most 64 characters.
- The seq monster: progress_seq was int4 while the flush ADVANCES it by
  arithmetic - overflow at INT_MAX (22003) errored the flush every tick
  and the terminal write's 22003 was misclassified as transient infra:
  the job stuck running, re-executed by every reclaim. Migrated to
  bigint (01.00.20_01) and the four deferral/release template casts
  follow. The JS-side 2^53 double-precision bound is documented at the
  comparison, and the poll's cursor reads the exact decimal from the
  ETag header (BigInt compare) instead of the mangled JSON double.
- The admin jobs list no longer selects error_message: nothing rendered
  it, and one poisoned row (a 10MB exception message) made every list
  view drag the whole value out of TOAST per row, per fetch.

Pins: tests/test_attack_input_monsters.py (the inventory), the realtime
harness stages the 2^53 collision and the 10k-deep data tick under Node.

* fix: lint/type discipline for the monster gates (ruff, pyright)

* chore: ignore the gremlins report artifact dir (coverage/)

* docs: the review fixes - the name cap documented as the breaking change it is, the migration's rewrite cost stated honestly, the depth cap's real bound, the gremlins gate made reproducible

- D2: docs/guides/upgrading.md gains the entry for the 255-char queue/
  actor name bound (an undocumented breaking change: on main a 300-char
  name passed and worked). The bound stays a CHARACTER count (choice a):
  the mechanism forces ~2700 BYTES per btree index item, but the item
  carries the actor name and the queue name SIDE BY SIDE, so a per-name
  bound has to compose; 255 chars is 1020 bytes at the worst-case 4
  bytes/char, and two of those plus the fixed-width tail fit one item.
  A byte-measured single-name bound would admit a queue name that only
  fails when paired with a legal actor name - the same opaque storage
  error the cap removes.
- D3: the migration header no longer claims "no index rebuild"; it states
  the full-table rewrite (fresh relfilenode, EVERY index rebuilt) and the
  ops window that scales with live row count. One pin holds the header to
  that story (no in-place / no-rebuild claim can return).
- D4: the depth-cap comment states the real bound (the cap sits BELOW the
  worker's acceptance: 200 deep stored, 600 refused) and the honest
  consequence (a below-cap-only change is dropped as a duplicate until
  something at or above the cap changes).
- D5: the gremlins gate is reproducible: committed [tool.pytest-gremlins]
  config naming the four touched modules, the killing suites in the
  documented command, --gremlin-executor=subprocess (the fork executor
  reports false kills: one run zapped 0/127). New pins close the gate's
  own survivor margins: the bounded echo's exact 64-char truncation edge,
  the offender's boundary and charset diagnoses. True numbers: 217
  mutants, 189/26 and 191/26 across two full runs, every survivor on a
  pre-existing line, 0 in the new gate code.

* chore: renumber the progress-seq-bigint migration to 01.00.20_03 - resolves the ledger-key collision with PR 513's watermark migration (490 keeps _01)

* docs(release-notes): mark the queue/actor name cap for the 0.3.0 notes

BREAKING CHANGE: queue and actor names are bounded at 255 characters: enqueueing to a queue name longer than 255 characters or declaring an actor name longer than 255 characters now raises ValueError at the chokepoint (the queue at enqueue, the actor at @actor registration), where on previous releases both were accepted and later failed the jobs INSERT with an opaque btree index-size error.

* test(release-notes): the 255-char name cap's upgrading.md section joins the breaking-change policy map

The completeness half of the guard fails while the guide's new Breaking
API changes section has no entry: the map exists to force docs
visibility, so the entry carries the same fields its siblings do - the
target release (0.3.0, still pending: the manifest is at 0.2.2) and the
needle the carrier must carry while pending (and the shipped bullet
after the cut). The needle's carrier is the preceding single-footer
marker commit, the structurally unambiguous form the parser surfaces.

+++ src/taskq/backend/_sql_templates.py
@@ -792,7 +792,7 @@ WITH params AS (
            -- never match neither).
             GREATEST($3::interval, {MIN_DEFERRAL_INTERVAL_SQL}) AS effective_delay,
             $4::jsonb AS metadata_update,
-            $5::int AS progress_seq,
+            $5::bigint AS progress_seq,
             $6::jsonb AS progress_state,
             $8::int AS attempt,
             -- The denial class the caller reported for THIS deferral (the
@@ -997,7 +997,7 @@ WITH params AS (
     SELECT $1::uuid AS job_id,
            $2::uuid AS worker_id,
            $3::interval AS delay,
-           $4::int AS progress_seq,
+           $4::bigint AS progress_seq,
            $5::jsonb AS progress_state,
            $6::int AS attempt,
            $7::bigint AS claim_epoch
@@ -1250,7 +1250,7 @@ WITH params AS (
            -- by construction, and it must never park the job at the
            -- head of the dispatch order either.
             GREATEST($3::interval, {MIN_DEFERRAL_INTERVAL_SQL}) AS effective_delay,
-            $4::int AS progress_seq,
+            $4::bigint AS progress_seq,
             $5::jsonb AS progress_state,
             $6::int AS attempt,
             $7::bigint AS claim_epoch
@@ -1468,7 +1468,7 @@ WITH params AS (
            CASE WHEN $4::interval > interval '0'
                 THEN GREATEST($4::interval, {MIN_DEFERRAL_INTERVAL_SQL})
                 ELSE interval '0' END AS effective_hold,
-           $5::int AS progress_seq,
+           $5::bigint AS progress_seq,
            $6::jsonb AS progress_state,
            $7::bigint AS claim_epoch
 ),
